IMS/requirements.txt
Django
django-cors-headers

djangorestframework
djangorestframework-simplejwt
django-versatileimagefield 3.1
python-magic-bin